django/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
celery==5.4.0
django-celery-results==2.5.0
gunicorn==22.0.0
psycopg[binary]==3.1.18
dj-database-url==2.1.0
pgvector==0.4.1
orjson>=3.10.3
pydantic>=2
plotly==5.22.0
torch==2.7.0
transformers==4.48.1
esm==3.2.0
biopython==1.85
tqdm==4.67.1
pyarrow
pandas
redis>=4.0.0
pyrodigal==3.6.3
umap-learn==0.5.7
numpy==2.0.1
joblib==1.4.2
seaborn==0.13.2
requests==2.32.3
django-filter==24.3
django-matomo==0.1.6
whitenoise==6.4.0
django-debug-toolbar==5.0.1
django-redis
pyhmmer==0.11.1
rdkit==2025.03.2
django-csp
uvicorn[standard]==0.30.1
aiofiles
python-multipart==0.0.9
django-ninja==1.4